+++ src/Faithlife.Build/BuildUtility.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using GlobExpressions;
 
@@ -14,17 +15,35 @@ namespace Faithlife.Build
 	public static class BuildUtility
 	{
 		/// <summary>
-		/// Finds the directories matching the specified globs.
+		/// Finds the files matching the specified globs, from the current working directory.
+		/// </summary>
+		/// <param name="globs">The globs to match.</param>
+		/// <returns>The paths of the matching files.</returns>
+		public static IReadOnlyList<string> FindFiles(params string[] globs) => FindFilesFrom(".", globs);
+
+		/// <summary>
+		/// Finds the directories matching the specified globs, from the current working directory.
 		/// </summary>
 		/// <param name="globs">The globs to match.</param>
 		/// <returns>The paths of the matching directories.</returns>
-		public static IReadOnlyList<string> FindDirectories(params string[] globs) => globs.SelectMany(glob => Glob.Directories(".", glob)).ToList();
+		public static IReadOnlyList<string> FindDirectories(params string[] globs) => FindDirectoriesFrom(".", globs);
 
 		/// <summary>
 		/// Finds the files matching the specified globs.
 		/// </summary>
+		/// <param name="directory">The starting directory.</param>
 		/// <param name="globs">The globs to match.</param>
 		/// <returns>The paths of the matching files.</returns>
-		public static IReadOnlyList<string> FindFiles(params string[] globs) => globs.SelectMany(glob => Glob.Files(".", glob)).ToList();
+		public static IReadOnlyList<string> FindFilesFrom(string directory, params string[] globs) =>
+			globs.SelectMany(glob => Glob.Files(directory, glob, GlobOptions.CaseInsensitive)).Distinct().Select(path => Path.Combine(directory, path)).ToList();
+
+		/// <summary>
+		/// Finds the directories matching the specified globs.
+		/// </summary>
+		/// <param name="directory">The starting directory.</param>
+		/// <param name="globs">The globs to match.</param>
+		/// <returns>The paths of the matching directories.</returns>
+		public static IReadOnlyList<string> FindDirectoriesFrom(string directory, params string[] globs) =>
+			globs.SelectMany(glob => Glob.Directories(directory, glob, GlobOptions.CaseInsensitive)).Distinct().Select(path => Path.Combine(directory, path)).ToList();
 	}
 }