+++ VersionHistory.md
@@ -6,6 +6,10 @@ Describe changes here when they're committed to the `master` branch. Move them t
 
 Prefix the description of the change with `[major]`, `[minor]`, or `[patch]` in accordance with [Semantic Versioning](https://semver.org/).
 
+* [minor] Add `FindFilesFrom` and `FindDirectoriesFrom` to `BuildUtility`.
+* [minor] Globs should use case-insensitive matching.
+* [patch] Don't return the same found file/directory twice.
+
 ## Released
 
 ### 0.5.0